feat: implement secure auth system with OAuth2, 2FA, and Redis

+++ src/main/java/com/weaver/weaver_backend/service/impl/AuthServiceImpl.java
@@ -13,12 +13,14 @@ import com.weaver.weaver_backend.entity.RedisToken;
 import com.weaver.weaver_backend.entity.User;
 import com.weaver.weaver_backend.exception.BadRequestException;
 import com.weaver.weaver_backend.exception.NotFoundException;
+import com.weaver.weaver_backend.exception.UnauthorizedException;
 import com.weaver.weaver_backend.mapper.UserMapper;
 import com.weaver.weaver_backend.repository.UserRepository;
 import com.weaver.weaver_backend.service.other.GoogleAuthenticatorService;
 import com.weaver.weaver_backend.service.IAuthService;
 import com.weaver.weaver_backend.service.IRedisTokenService;
 import com.weaver.weaver_backend.util.JwtUtils;
+import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -50,31 +52,18 @@ public class AuthServiceImpl implements IAuthService {
         if (!passwordEncoder.matches(password, user.getPassword())) {
             throw new BadRequestException("Invalid email or password");
         }
-        if (user.getTwoFaEnabled()) {
-            TokenResponse twoFAToken = jwtUtils.generateToken(user, TokenType.TWOFA_TOKEN);
-            return LoginResponse.builder()
-                    .twoFAToken(twoFAToken.value())
-                    .build();
-        }
-        TokenResponse accessToken = jwtUtils.generateToken(user, TokenType.ACCESS_TOKEN);
-        TokenResponse refreshToken = jwtUtils.generateToken(user, TokenType.REFRESH_TOKEN);
-        RedisToken redisToken = RedisToken.builder()
-                .jwtId(refreshToken.jwtId())
-                .userId(user.getId())
-                .expiration(refreshToken.ttlSeconds())
-                .build();
-
-        redisTokenService.saveToken(redisToken);
-        return LoginResponse.builder()
-                .accessToken(accessToken.value())
-                .refreshToken(refreshToken.value())
-                .build();
+        return handleLoginSuccess(user);
     }
 
     @Override
     public LoginResponse verifyTwoFA(String token, int OTP) {
         try {
-            UUID userId = jwtUtils.getUserId(token);
+            Claims claims = jwtUtils.extractClaims(token);
+            UUID userId = jwtUtils.getUserId(claims);
+            TokenType tokenType = jwtUtils.getType(claims);
+            if (tokenType != TokenType.TWOFA_TOKEN) {
+                throw new BadRequestException("Invalid 2FA token");
+            }
             User user = userRepository.findById(userId)
                     .orElseThrow(() -> new NotFoundException("User not found"));
             boolean isVerified = ggAuthService.verifyCode(user.getTwoFaSecret(), OTP);
@@ -121,23 +110,23 @@ public class AuthServiceImpl implements IAuthService {
                     return userRepository.save(newUser);
                 }
         );
-        TokenResponse accessToken = jwtUtils.generateToken(user, TokenType.ACCESS_TOKEN);
-        TokenResponse refreshToken = jwtUtils.generateToken(user, TokenType.REFRESH_TOKEN);
-        RedisToken redisToken = RedisToken.builder()
-                .jwtId(refreshToken.jwtId())
-                .userId(user.getId())
-                .expiration(refreshToken.ttlSeconds())
-                .build();
-        redisTokenService.saveToken(redisToken);
-        return LoginResponse.builder()
-                .accessToken(accessToken.value())
-                .refreshToken(refreshToken.value())
-                .build();
+        return handleLoginSuccess(user);
     }
 
     @Override
     public LoginResponse refreshToken(String refreshToken) {
-        UUID userId = jwtUtils.getUserId(refreshToken);
+        Claims refreshClaims = jwtUtils.extractClaims(refreshToken);
+        TokenType tokenType = jwtUtils.getType(refreshClaims);
+        UUID userId = jwtUtils.getUserId(refreshClaims);
+        String refreshJwtId = jwtUtils.getJwtId(refreshClaims);
+
+        if (tokenType != TokenType.REFRESH_TOKEN) {
+            throw new BadRequestException("Invalid refresh token");
+        }
+
+        if (!redisTokenService.existsByJwtId(refreshJwtId)) {
+            throw new UnauthorizedException("Refresh token has been revoked or logged out");
+        }
 
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("User not found"));
@@ -155,15 +144,22 @@ public class AuthServiceImpl implements IAuthService {
         if (refreshToken == null || refreshToken.isEmpty()) {
             throw new BadRequestException("Missing logout information");
         }
-        String refreshJwtId = jwtUtils.getJwtId(refreshToken);
+        Claims refreshClaims = jwtUtils.extractClaims(refreshToken);
+        TokenType tokenType = jwtUtils.getType(refreshClaims);
+        String refreshJwtId = jwtUtils.getJwtId(refreshClaims);
+        if (tokenType != TokenType.REFRESH_TOKEN) {
+            throw new BadRequestException("Invalid refresh token");
+        }
+        log.info("Refresh token {} deleted from Redis", refreshJwtId);
         redisTokenService.deleteTokenByJwtId(refreshJwtId);
         String header = httpServletRequest.getHeader("Authorization");
         if (header != null && header.startsWith("Bearer ")) {
             String accessToken = header.substring(7);
-            Date exp = jwtUtils.getExpiration(accessToken);
-            long ttl = exp.getTime() - System.currentTimeMillis();
-            String accessJwtId = jwtUtils.getJwtId(accessToken);
-            UUID userId = jwtUtils.getUserId(accessToken);
+            Claims accessClaims = jwtUtils.extractClaims(accessToken);
+            Date exp = jwtUtils.getExpiration(accessClaims);
+            long ttl = (exp.getTime() - System.currentTimeMillis()) / 1000;
+            String accessJwtId = jwtUtils.getJwtId(accessClaims);
+            UUID userId = jwtUtils.getUserId(accessClaims);
             if (ttl > 0) {
                 redisTokenService.saveToken(
                         RedisToken.builder()
@@ -172,13 +168,38 @@ public class AuthServiceImpl implements IAuthService {
                                 .expiration(ttl)
                                 .build()
                 );
-                log.info("Access token blacklisted, TTL: {} ms", ttl);
+                log.info("Access token {} blacklisted for {} ms", accessJwtId, ttl);
             } else {
                 log.info("Access token already expired (no need to blacklist)");
             }
         } else {
             log.warn("No Bearer token found in Authorization header, skipping blacklist");
+            throw new UnauthorizedException("Please login");
         }
         SecurityContextHolder.clearContext();
     }
+
+    private LoginResponse handleLoginSuccess(User user) {
+        if (user.getTwoFaEnabled()) {
+            TokenResponse twoFAToken = jwtUtils.generateToken(user, TokenType.TWOFA_TOKEN);
+            return LoginResponse.builder()
+                    .twoFAToken(twoFAToken.value())
+                    .build();
+        }
+
+        TokenResponse accessToken = jwtUtils.generateToken(user, TokenType.ACCESS_TOKEN);
+        TokenResponse refreshToken = jwtUtils.generateToken(user, TokenType.REFRESH_TOKEN);
+
+        redisTokenService.saveToken(RedisToken.builder()
+                .jwtId(refreshToken.jwtId())
+                .userId(user.getId())
+                .expiration(refreshToken.ttlSeconds())
+                .build());
+
+        return LoginResponse.builder()
+                .accessToken(accessToken.value())
+                .refreshToken(refreshToken.value())
+                .build();
+    }
+
 }
